refactor: 바뀐 구조에 따른 패키지 분리
