+++ MyMate/app/src/main/AndroidManifest.xml
@@ -184,7 +184,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name=".presentation.main.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.MyMate">
             <intent-filter>
