+++ MyMate/app/src/main/java/com/example/mymate/MainReportMeFragment.kt
@@ -23,6 +23,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.example.mymate.data.dto.common.DefaultResponse
 import com.example.mymate.data.dto.report.response.UserReportResponse
 import com.example.mymate.databinding.MainReportMeFragmentBinding
+import com.example.mymate.presentation.main.MainActivity
 import com.github.mikephil.charting.data.PieData
 import com.github.mikephil.charting.data.PieDataSet
 import com.github.mikephil.charting.data.PieEntry
