+++ MyMate/app/src/main/res/layout/main_home_fragment.xml
@@ -3,7 +3,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         xmlns:tools="http://schemas.android.com/tools">
     <data>
-        <variable name="vm" type="com.example.mymate.MainHomeViewModel"/>
+        <variable name="vm" type="com.example.mymate.presentation.home.MainHomeViewModel"/>
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
